pyproject.toml
[tool.poetry]
name = "python-django-module-2"
version = "0.1.0"
description = ""
authors = ["Александр Помаз <aramissasha@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
django-rest-framework = "^0.1.0"
asgiref = "3.7.2"
charset-normalizer = "3.3.2"
django-debug-toolbar = "4.2.0"
django-dumpdata-field = "0.1"
attrs = "23.2.0"
django-filter = "23.5"
djangorestframework = "3.14.0"
docker = "7.0.0"
docutils = "0.20.1"
drf-spectacular = "0.27.0"
flake8 = "7.0.0"
flake8-docstrings = "1.7.0"
gunicorn = "21.2.0"
idna = "3.6"
inflection = "0.5.1"
jsonschema = "4.21.0"
jsonschema-specifications = "2023.12.1"
mccabe = "0.7.0"
packaging = "23.2"
pillow = "10.2.0"
pycodestyle = "2.11.1"
pydocstyle = "6.3.0"
pyflakes = "3.2.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
referencing = "0.32.1"
requests = "2.31.0"
rpds-py = "0.17.1"
sentry-sdk = "1.39.2"
snowballstemmer = "2.2.0"
sqlparse = "0.4.4"
uritemplate = "4.1.1"
urllib3 = "2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
